refactor: simplify CLI and move AI tools into backend package
