+++ docs/README.md
@@ -26,6 +26,7 @@ Always check `nsys --version` on the target system and cross-reference with the 
 | [06-python-pytorch.md](./06-python-pytorch.md) | Python and PyTorch profiling support | User Guide |
 | [07-container-profiling.md](./07-container-profiling.md) | Docker/container profiling setup | User Guide |
 | [08-focused-profiling.md](./08-focused-profiling.md) | Limiting profile scope with cudaProfilerApi and NVTX capture ranges | User Guide |
+| [09-performance-questions-mfu.html](./09-performance-questions-mfu.html) | Curated high-impact performance questions and MFU calculation playbook | Project Guide |
 
 ## How This Knowledge Base Is Used
