+++ src/nsys_ai/ai/backend/__init__.py
@@ -1,0 +1,27 @@
+"""Backend internals for AI tool orchestration."""
+
+from .chat_tools import _build_system_prompt, _parse_tool_call, _tools_openai
+from .profile_db_tool import (
+    DEFAULT_MAX_JSON_CHARS,
+    DEFAULT_MAX_LIMIT,
+    TOOL_QUERY_PROFILE_DB,
+    get_profile_schema,
+    get_profile_schema_cached,
+    open_profile_readonly,
+    open_profile_readonly_for_worker,
+    query_profile_db,
+)
+
+__all__ = [
+    "_build_system_prompt",
+    "_parse_tool_call",
+    "_tools_openai",
+    "DEFAULT_MAX_JSON_CHARS",
+    "DEFAULT_MAX_LIMIT",
+    "TOOL_QUERY_PROFILE_DB",
+    "get_profile_schema",
+    "get_profile_schema_cached",
+    "open_profile_readonly",
+    "open_profile_readonly_for_worker",
+    "query_profile_db",
+]
