+++ src/nsys_ai/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI package for nsys-ai."""
+
+from .app import main
+
+__all__ = ["main"]
